+++ crates/nostr-relay-builder/src/builder.rs
@@ -5,6 +5,10 @@
 //! Relay Builder
 
 use std::net::IpAddr;
+#[cfg(all(feature = "tor", any(target_os = "android", target_os = "ios")))]
+use std::path::Path;
+#[cfg(feature = "tor")]
+use std::path::PathBuf;
 use std::sync::Arc;
 
 use nostr_database::prelude::*;
@@ -28,6 +32,53 @@ impl Default for RateLimit {
     }
 }
 
+/// Relay builder tor hidden service options
+#[cfg(feature = "tor")]
+pub struct RelayBuilderHiddenService {
+    /// Nickname (local identifier) for a Tor hidden service
+    ///
+    /// Used to look up this service's keys, state, configuration, etc., and distinguish them from other services.
+    pub nickname: String,
+    /// Custom path
+    pub custom_path: Option<PathBuf>,
+}
+
+#[cfg(feature = "tor")]
+impl RelayBuilderHiddenService {
+    /// New tor hidden service options
+    ///
+    /// The nickname is a local identifier for a Tor hidden service.
+    /// It's used to look up this service's keys, state, configuration, etc., and distinguish them from other services.
+    #[inline]
+    #[cfg(all(not(target_os = "android"), not(target_os = "ios")))]
+    pub fn new<S>(nickname: S) -> Self
+    where
+        S: Into<String>,
+    {
+        Self {
+            nickname: nickname.into(),
+            custom_path: None,
+        }
+    }
+
+    /// New tor hidden service options
+    ///
+    /// The nickname is a local identifier for a Tor hidden service.
+    /// It's used to look up this service's keys, state, configuration, etc., and distinguish them from other services.
+    #[inline]
+    #[cfg(any(target_os = "android", target_os = "ios"))]
+    pub fn new<S, P>(nickname: S, path: P) -> Self
+    where
+        S: Into<String>,
+        P: AsRef<Path>,
+    {
+        Self {
+            nickname: nickname.into(),
+            custom_path: Some(path.as_ref().to_path_buf()),
+        }
+    }
+}
+
 /// Relay builder
 pub struct RelayBuilder {
     /// IP address
@@ -38,6 +89,9 @@ pub struct RelayBuilder {
     pub database: Arc<DynNostrDatabase>,
     /// Rate limit
     pub rate_limit: RateLimit,
+    /// Tor hidden service
+    #[cfg(feature = "tor")]
+    pub tor: Option<RelayBuilderHiddenService>,
 }
 
 impl Default for RelayBuilder {
@@ -50,6 +104,8 @@ impl Default for RelayBuilder {
                 max_events: Some(75_000),
             })),
             rate_limit: RateLimit::default(),
+            #[cfg(feature = "tor")]
+            tor: None,
         }
     }
 }
@@ -83,6 +139,14 @@ impl RelayBuilder {
     #[inline]
     pub fn rate_limit(mut self, limit: RateLimit) -> Self {
         self.rate_limit = limit;
+        self
+    }
+
+    /// Set tor options
+    #[inline]
+    #[cfg(feature = "tor")]
+    pub fn tor(mut self, opts: RelayBuilderHiddenService) -> Self {
+        self.tor = Some(opts);
         self
     }
 }