+++ crates/nostr-relay-builder/src/error.rs
@@ -6,6 +6,8 @@
 
 use std::io;
 
+#[cfg(feature = "tor")]
+use async_wsocket::native::tor;
 use thiserror::Error;
 
 /// Relay builder error
@@ -14,6 +16,10 @@ pub enum Error {
     /// I/O error
     #[error(transparent)]
     IO(#[from] io::Error),
+    /// Tor error
+    #[cfg(feature = "tor")]
+    #[error(transparent)]
+    Tor(#[from] tor::Error),
     /// No port available
     #[error("No port available")]
     NoPortAvailable,