+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 ### Added
 
 * relay-builder: add `LocalRelay` and `RelayBuilder` ([Yuki Kishimoto])
+* relay-builder: allow to serve local relay as hidden service ([Yuki Kishimoto])
 * pool: add `RelayPool::force_remove_relay` method ([Yuki Kishimoto])
 * zapper: add `WebLNZapper` struct (moved from `nostr-webln` crate) ([Yuki Kishimoto])